refactor tron api
